+++ crates/astria-conductor/src/test_utils.rs
@@ -1,5 +1,5 @@
 use astria_core::{
-    generated::astria::execution::v1::{
+    generated::astria::execution::v2::{
         CommitmentState,
         GenesisInfo,
     },
@@ -10,7 +10,7 @@ use astria_core::{
 use crate::executor::State;
 
 pub(crate) fn make_commitment_state() -> CommitmentState {
-    let firm = astria_core::generated::astria::execution::v1::Block {
+    let firm = astria_core::generated::astria::execution::v2::Block {
         number: 1,
         hash: vec![42u8; 32].into(),
         parent_block_hash: vec![41u8; 32].into(),
@@ -19,7 +19,7 @@ pub(crate) fn make_commitment_state() -> CommitmentState {
             nanos: 789,
         }),
     };
-    let soft = astria_core::generated::astria::execution::v1::Block {
+    let soft = astria_core::generated::astria::execution::v2::Block {
         number: 2,
         hash: vec![43u8; 32].into(),
         parent_block_hash: vec![42u8; 32].into(),
@@ -54,8 +54,8 @@ pub(crate) fn make_rollup_state(
     genesis_info: GenesisInfo,
     commitment_state: CommitmentState,
 ) -> State {
-    let genesis_info = astria_core::execution::v1::GenesisInfo::try_from_raw(genesis_info).unwrap();
+    let genesis_info = astria_core::execution::v2::GenesisInfo::try_from_raw(genesis_info).unwrap();
     let commitment_state =
-        astria_core::execution::v1::CommitmentState::try_from_raw(commitment_state).unwrap();
+        astria_core::execution::v2::CommitmentState::try_from_raw(commitment_state).unwrap();
     State::try_from_genesis_info_and_commitment_state(genesis_info, commitment_state).unwrap()
 }